pyproject.toml
[project]
name = "fisherA2Z"
version = "0.3.0"
authors = [
  { name="Husni Almoubayyed" },
  { name="Tianqing Zhang", email = "tq96zhang@gmail.com"}]
description = "A Fisher forecasting code for the LSST 3x2pt analysis"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pandas",
    "pyccl",
    "numpy<2.0.0",
    "matplotlib",
    "numdifftools==0.9.39",
    "scipy",
    "scikit-learn==1.1.0",
    "PyYAML",
    "jupyter",
    "qp-prob",
    "h5py"
]

[project.urls]
Homepage = "https://github.com/LSSTDESC/fisherA2Z"
Issues = "https://github.com/LSSTDESC/fisherA2Z/issues"

[project.scripts]
forecast-pz = "fisherA2Z.cli:main"

[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8", "cmake", "swig"]
build-backend = "setuptools.build_meta"